Group files: logic vs components
