+++ src/components/Card.css
@@ -1,0 +1,8 @@
+.Card {
+  height: 125px;
+  width: 100px;
+
+  border: 5px solid transparent;
+  margin: 1px;
+  padding: 1px;
+}
