+++ src/components/Card.js
@@ -1,11 +1,12 @@
 import React from "react";
+import './Card.css';
 
 export default function Card({ suit, value, color }) {
   return (
     <img
       src={`http://h3h.net/images/cards/${suit}_${value}.svg`}
       alt={`${value} of ${suit}`}
-      className="card"
+      className="Card"
       style={{ borderColor: color }}
     />
   );
