+++ src/components/Footer.css
@@ -1,0 +1,34 @@
+/* Add/remove hand button */
+.Footer .change-hands-number-button {
+  background-color: #4b1124;
+  border-width: 0px;
+  box-shadow: 4px 4px 5px #333;
+  color: #eee;
+  cursor: pointer;
+  font-size: 24px;
+  padding: 0.5em 2em;
+  margin: 5px;
+}
+
+.Footer .change-hands-number-button:disabled {
+  cursor: not-allowed;
+  color: gray;
+  background-color: #5a3642;
+}
+
+/* Play button */
+.Footer .play-button {
+  background-color: #4b1124;
+  border: 4px double #cfb338;
+  box-shadow: 4px 4px 5px #333;
+  color: #eee;
+  cursor: pointer;
+  font-size: 24px;
+  padding: 0.5em 2em;
+}
+.Footer .play-button:active {
+  position: relative;
+  top: 4px;
+  left: 4px;
+  box-shadow: none;
+}
