+++ src/components/Hand.css
@@ -1,0 +1,3 @@
+.Hand {
+  text-align: center;
+}
