+++ src/components/Hand.js
@@ -1,7 +1,7 @@
 import React from "react";
-import Card from "../card";
+import Card from "./Card";
 
-import './hand.css';
+import './Hand.css';
 
 export default function Hand({ cards, colors }) {
   // We might want to define the coloring for cards in a more sophisticated way, but it should be fine for now
