+++ src/components/Player.css
@@ -1,0 +1,15 @@
+.Player {
+    text-align: center;
+    margin: 20px 0;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+}
+
+.Player .playerName, .Player .pairs {
+    width: 200px;
+}
+
+.Player .Hand {
+    flex-grow: 1;
+}
