+++ src/components/index.js
@@ -1,0 +1,6 @@
+import Card from "./Card";
+import Footer from "./Footer";
+import Hand from "./Hand";
+import Player from "./Player";
+
+export { Card, Footer, Hand, Player };
