--- src/hand/hand.css
@@ -1,3 +1,0 @@
-.Hand {
-    text-align: center;
-}
