+++ src/index.css
@@ -17,65 +17,3 @@ h1 {
   text-shadow: 1px 2px #444;
 }
 
-/* Cards */
-.card {
-  height: 125px;
-  width: 100px;
-
-  border: 5px solid transparent;
-  margin: 1px;
-  padding: 1px;
-}
-
-.winnerString {
-  text-align: center;
-}
-
-/* Add/remove hand button */
-.change-hands-number-button {
-  background-color: #4b1124;
-  border-width: 0px;
-  box-shadow: 4px 4px 5px #333;
-  color: #eee;
-  cursor: pointer;
-  font-size: 24px;
-  padding: 0.5em 2em;
-  margin: 5px;
-}
-
-.change-hands-number-button:disabled {
-  cursor: not-allowed;
-  color: gray;
-  background-color: #5a3642;
-}
-
-/* Button */
-.play-button {
-  background-color: #4b1124;
-  border: 4px double #cfb338;
-  box-shadow: 4px 4px 5px #333;
-  color: #eee;
-  cursor: pointer;
-  font-size: 24px;
-  padding: 0.5em 2em;
-}
-.play-button:active {
-  position: relative;
-  top: 4px;
-  left: 4px;
-  box-shadow: none;
-}
-
-.App {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-}
-main {
-    flex: 1 0 auto;
-}
-footer {
-  flex-shrink: 0;
-  text-align: center;
-  padding-bottom: 30px;
-}
