+++ src/logic/constants.js
@@ -1,0 +1,6 @@
+export const maxHands = 4;
+export const minHands = 2;
+export const initialHands = 2;
+
+// NOTE: If we ever want to increase the number of cards in hand, we'd have to add more colors to the colors array
+export const cardsPerHand = 7;
