--- src/player/player.css
@@ -1,4 +1,0 @@
-.Player {
-    text-align: center;
-    margin-bottom: 40px;
-}
